Added req/res interactions for recipe creation

+++ server.js
@@ -10,6 +10,7 @@ const app = express();
 
 app.use(cors());
 app.use(express.json());
+app.use(express.urlencoded({ extended: false }))  // Lets us see information in the body
 
 
 /**
@@ -46,20 +47,15 @@ app.get("/recipe/:id", (request, response) => {
 // Create a new recipe
 // Send in a request where the request body has recipe information
 app.post("/recipe", (request, response) => {
-   
-    // Find a valid filename
-    var fileBasename = request.body.name;
-    var fileBasenameSuffix = ""
-    var numberSuffix = 0;
-    if(!fileBasename){
-        fileBasename = "unnamed";
+
+    if(!request.body.name){
+        return response.status(400).send({"error": "No recipe name specified."});
     }
-    var filepath = path.join("serverFiles", "recipes", `${fileBasename}${fileBasenameSuffix}.json`)
-    while(fs.existsSync(filepath)){
+
+    var filepath = path.join("serverFiles", "recipes", `${request.body.name}.json`);
+    if(fs.existsSync(filepath)){
         // Stops from overwriting a pre-existing file
-        numberSuffix += 1;
-        fileBasenameSuffix = String(numberSuffix);
-        filepath = path.join("serverFiles", "recipes", `${fileBasename}${fileBasenameSuffix}.json`);
+        return response.status(400).send({"error": "Recipe name already taken."});
     }
 
     fs.writeFile(filepath, JSON.stringify(request.body, null, "\t"), err => {
@@ -75,6 +71,23 @@ app.post("/recipe", (request, response) => {
         return;
     });
 });
+
+// Communicates if a recipe name is already taken
+// Body must have {"name": recipeName}
+app.get("/recipe-exists", (request, response) => {
+
+    if(!request.query.name){
+        return response.status(400).send({"error": "No recipe name specified."});
+    }
+
+    var fileExists = false;
+    var filepath = path.join("serverFiles", "recipes", `${request.query.name}.json`);
+    if(fs.existsSync(filepath)){
+        fileExists = true;
+    }
+
+    return response.status(200).send({"exists": fileExists});
+})
 
 app.put("/recipe/:id", (request, response) => {
     fs.writeFile(`./serverFiles/recipes/${request.params.id}`, err => {
